+++ src/servers/PrerenderServer.ts
@@ -2,6 +2,6 @@ export type PrerenderServer = {
     isServerReady(): Promise<void>
     destroy(): void
     get baseUrl(): string
-    get staticDir(): string
+    get publicDir(): string
     get publicPath(): string
 }